+++ FoundryReports.Core/Products/Mold.cs
@@ -5,9 +5,9 @@ namespace FoundryReports.Core.Products
     [Serializable]
     internal class Mold : IMold
     {
-        public int MaxUsages { get; set;  }
+        public decimal MaxUsages { get; set;  }
 
-        public int CurrentUsages { get; set; }
+        public decimal CurrentUsages { get; set; }
 
         public decimal CastingCellAmount { get; set; }
