+++ FoundryReports.Core/Products/MoldDummy.cs
@@ -2,8 +2,8 @@
 {
     internal class MoldDummy : IMold
     {
-        public int MaxUsages { get; set; }
-        public int CurrentUsages { get; set; }
+        public decimal MaxUsages { get; set; }
+        public decimal CurrentUsages { get; set; }
         public decimal CastingCellAmount { get; set; }
         public string Name { get; set; } = string.Empty;
     }
